fix method for adding truck for driver

+++ src/main/java/org/example/logisticapplication/domain/Truck/TruckDto.java
@@ -18,9 +18,9 @@ public record TruckDto(
 
         Double capacity,
 
-        Integer currentCityId,
+        Long currentCityId,
 
-        Integer currentDriverId
+        Long currentDriverId
 
 ) {
 }
